Memory Defragmentation
Consider an analogue of the computer's memory on the example of the row array. If you delete some items, then the so-called "holes" will appear in the array - that is, elements are equal to NULL. The ExeCuedEfragmentation method (String []) accepting an array of strings performs it "defragmentation", that is, moves all objects to the beginning of the array in the same order by moving all the "holes" (elements equal to NULL) to the end of the array. In this task you need to implement the EXECUTEDEFRAGMENTATION method (String []). Main () method You can use to verify the result of the executedefragmentation method (String [])